requirements.txt
# ── LLM & Agents (depuis Agent Cyber) ────────────────────────
openai-agents
python-dotenv

# ── RAG Pipeline (depuis RAG Cybersec) ───────────────────────
langchain-classic>=0.0.1
langchain-community>=0.3.0
langchain-chroma>=0.1.4
langchain-huggingface>=0.1.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
rank-bm25>=0.2.2
sentence-transformers>=3.0.0
pypdf>=4.0.0

# ── MCP ──────────────────────────────────────────────────────
fastmcp
composio-openai-agents

# ── API & Web ────────────────────────────────────────────────
requests
pydantic>=2.0
streamlit-oauth>=0.1.14
# ── Frontend ─────────────────────────────────────────────────
streamlit
